git-dsc-commit: remove --skip-patches flag
While the original behavior of this tools was to remove
applied patches after building the tree, the current usage
needs the patches to be applied. So restoring the original
behavior and removing --skip-patches is necessary.

Also, the previous code to remove the patches after the tree
was created will be kept out